Move useEffects and associated functions to HomeControl from Home, general cleanup
